+++ src/test/java/com/itransition/payment/AssertionsHelper.java
@@ -1,4 +1,4 @@
-package com.itransition.payment.core;
+package com.itransition.payment;
 
 import com.itransition.payment.transaction.entity.Transaction;
 import com.itransition.payment.account.dto.AccountDto;
